+++ compiler/src/iree/compiler/Codegen/Common/GPUReduceBankConflicts.cpp
@@ -61,7 +61,8 @@ struct GPUReduceBankConflictsPass
     // Collect all the alloc operations.
     funcOp.walk([&](memref::AllocOp allocOp) {
       if (allocOp.getType().getMemorySpaceAsInt() ==
-              gpu::GPUDialect::getWorkgroupAddressSpace() &&
+              static_cast<unsigned int>(
+                  gpu::GPUDialect::getWorkgroupAddressSpace()) &&
           allocOp.getType().hasStaticShape()) {
         sharedMemAllocs.push_back(allocOp);
       }
